pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "google-genai==1.0.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-mongodb==0.6.2",
    "langchain-openai>=0.3.28",
    "motor<3.6",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.10.1",
    "pymongo<4.9",
    "python-dotenv==1.0.0",
    "python-socketio[asyncio-client]==5.9.0",
    "uvicorn[standard]==0.24.0",
    "websockets>=14.2",
]
